requirements.txt
beautifulsoup4==4.13.4
EbookLib==0.19
mutagen==1.47.0
openai==1.85.0
requests==2.32.4
edge-tts==7.0.2
pydub==0.25.1
wyoming==1.6.0
gradio==5.33.1
docker==7.1.0
sentencex==0.6.1
gradio_log==0.0.8
pystray==0.19.5
Pillow==10.4.0
python-dotenv==1.0.0
num2words==0.5.13
TTS